+++ docs/changelog.rst
@@ -22,6 +22,7 @@ Fixes
 
 * Fix some metadata in magnettech importer (experiment.runs -> signal_channel.accumulations, correctly import spectrometer metadata, bring time stamp to same timezone.)
 
+* Do range extraction and interpolation (instead of interpolation only) in GoniometerSweepImporter.
 
 
 Version 0.3.0